Move the passphraseRetriever to its own file.
