+++ cmd/notary-signer/config.go
@@ -1,3 +1,16 @@
 package main
 
-import ()
+import (
+	"errors"
+	"strings"
+)
+
+func passphraseRetriever(keyName, alias string, createNew bool, attempts int) (passphrase string, giveup bool, err error) {
+	passphrase = mainViper.GetString(strings.ToUpper(alias))
+
+	if passphrase == "" {
+		return "", false, errors.New("expected env variable to not be empty: " + alias)
+	}
+
+	return passphrase, false, nil
+}
